build: separate backend and project-loader versions
Versions are now given in gradle.properties of the root project.
