+++ buildSrc/src/main/kotlin/backend-conventions.gradle.kts
@@ -1,0 +1,7 @@
+import de.itemis.mps.buildbackends.computeVersionSuffix
+
+plugins {
+    id("kotlin-conventions")
+}
+
+version = "${project.extra["version.backend"]}${computeVersionSuffix()}"
