+++ gradle.properties
@@ -1,0 +1,2 @@
+version.backend=1.10
+version.project-loader=1.10
